Implement forms + some other stuff - Patch API Controller to work with plain old HTML forms - Main view and forms ready - list view still missing - Add support for implementing JavaScript for the application

+++ src/main/java/com/example/student/dao/StudentJDBCTemplate.java
@@ -118,7 +118,7 @@ public class StudentJDBCTemplate implements StudentDao {
             }
         }, keyHolder);
 
-        Student student = new Student((Integer)keyHolder.getKey().intValue(), name, gpax, ambition);
+        Student student = new Student(keyHolder.getKey().intValue(), name, gpax, ambition);
         return student;
     }
 }
